+++ sklearn/externals/joblib/_compat.py
@@ -4,6 +4,8 @@ Compatibility layer for Python 3/Python 2 single codebase
 import sys
 
 PY3_OR_LATER = sys.version_info[0] >= 3
+PY26 = sys.version_info[:2] == (2, 6)
+PY27 = sys.version_info[:2] == (2, 7)
 
 try:
     _basestring = basestring
@@ -11,3 +13,8 @@ try:
 except NameError:
     _basestring = str
     _bytes_or_unicode = (bytes, str)
+
+
+def with_metaclass(meta, *bases):
+    """Create a base class with a metaclass."""
+    return meta("NewBase", bases, {})